+++ glwidget.h
@@ -19,7 +19,7 @@ private:
     Model* myModel;
     void createAxisPaint();
     void createRobotFrame();
-    void drawBox(double x1, double y1,double z1, double x2, double y2, double z2);
+    void drawBox(float x1, float y1,float z1, float x2, float y2, float z2);
 };
 
 #endif // GLWIDGET_H
